Apollova-Installer/requirements.txt
# Apollova - Requirements
# Install with: pip install -r requirements.txt

# Core
pytubefix>=6.0.0
pydub>=0.25.1
librosa>=0.10.0
requests>=2.28.0

# Optional (for .env support)
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0
colorthief>=0.2.1

# Lyrics processing
beautifulsoup4>=4.12.0
rapidfuzz>=3.0.0
stable-whisper>=2.15.0

# NOTE: FFmpeg must be installed separately
# Windows: choco install ffmpeg
# Or download from: https://ffmpeg.org/download.html
